https://github.com/krasserm/akka-persistence-cassandra/issues/96 initial commit. Write side journal table definition.
