+++ src/main/scala/akka/persistence/cassandra/old/CassandraJournal.scala
@@ -1,4 +1,5 @@
-package akka.persistence.cassandra.journal
+/*
+package akka.persistence.cassandra.old
 
 import java.lang.{Long => JLong}
 import java.nio.ByteBuffer
@@ -6,10 +7,12 @@ import java.nio.ByteBuffer
 import scala.collection.JavaConversions._
 import scala.collection.immutable.Seq
 import scala.concurrent._
+import scala.math.min
 import scala.util.{Failure, Success, Try}
 
 import akka.persistence._
 import akka.persistence.cassandra._
+import akka.persistence.cassandra.journal.CassandraJournalConfig
 import akka.persistence.journal.AsyncWriteJournal
 import akka.serialization.SerializationExtension
 import com.datastax.driver.core._
@@ -18,15 +21,6 @@ import com.datastax.driver.core.policies.{LoggingRetryPolicy, RetryPolicy}
 import com.datastax.driver.core.utils.Bytes
 
 class CassandraJournal extends AsyncWriteJournal with CassandraRecovery with CassandraStatements {
-
-  // TODO: journalId management.
-  // TODO: Cluster membership can change and new Journal instances may be added and old removed.
-  // TODO: We need to ensure globally unique journalId. Conflicts would violate the single writer requirement.
-  // TODO: Garbage collecting or infinitely growing journalId set?
-  private[this] val journalId = context.self.path.toString
-  println(journalId)
-
-  private[this] var journalSequenceNr = 0L
 
   val config = new CassandraJournalConfig(context.system.settings.config.getConfig("cassandra-journal"))
   val serialization = SerializationExtension(context.system)
@@ -56,28 +50,28 @@ class CassandraJournal extends AsyncWriteJournal with CassandraRecovery with Cas
   session.execute(writeConfig, CassandraJournalConfig.TargetPartitionProperty, config.targetPartitionSize.toString)
 
   val preparedWriteMessage = session.prepare(writeMessage)
+  val preparedDeletePermanent = session.prepare(deleteMessage)
+  val preparedSelectMessages = session.prepare(selectMessages).setConsistencyLevel(readConsistency)
+  val preparedCheckInUse = session.prepare(selectInUse).setConsistencyLevel(readConsistency)
   val preparedWriteInUse = session.prepare(writeInUse)
+  val preparedSelectHighestSequenceNr = session.prepare(selectHighestSequenceNr).setConsistencyLevel(readConsistency)
+  val preparedSelectDeletedTo = session.prepare(selectDeletedTo).setConsistencyLevel(readConsistency)
+  val preparedInsertDeletedTo = session.prepare(insertDeletedTo).setConsistencyLevel(writeConsistency)
 
-  override def asyncWriteMessages(messages: Seq[AtomicWrite]): Future[Seq[Try[Unit]]] = {
+  def asyncWriteMessages(messages: Seq[AtomicWrite]): Future[Seq[Try[Unit]]] = {
     // we need to preserve the order / size of this sequence even though we don't map
     // AtomicWrites 1:1 with a C* insert
-    val newJournalSequenceNr = journalSequenceNr + messages.size
-
-    val serialized = (journalSequenceNr to newJournalSequenceNr)
-      .zip(messages)
-      .map(aw => Try { SerializedAtomicWrite(
-        aw._2.payload.head.persistenceId,
-        aw._2.payload.map(pr => Serialized(aw._1, pr.sequenceNr, persistentToByteBuffer(pr))))
-      })
-    journalSequenceNr = newJournalSequenceNr
-
+    val serialized = messages.map(aw => Try { SerializedAtomicWrite(
+        aw.payload.head.persistenceId,
+        aw.payload.map(pr => Serialized(pr.sequenceNr, persistentToByteBuffer(pr))))
+    })
     val result = serialized.map(a => a.map(_ => ()))
 
     val byPersistenceId = serialized.collect({ case Success(caw) => caw }).groupBy(_.persistenceId).values
     val boundStatements = byPersistenceId.map(statementGroup)
 
     val batchStatements = boundStatements.map({ unit =>
-      executeBatch(batch => unit.foreach(batch.add))
+        executeBatch(batch => unit.foreach(batch.add))
     })
     val promise = Promise[Seq[Try[Unit]]]()
 
@@ -90,12 +84,9 @@ class CassandraJournal extends AsyncWriteJournal with CassandraRecovery with Cas
   }
 
   private def statementGroup(atomicWrites: Seq[SerializedAtomicWrite]): Seq[BoundStatement] = {
-    val firstJournalSequenceNr = atomicWrites.last.payload.last.journaSequenceNr
-    val lastJournalSequenceNr = atomicWrites.head.payload.head.journaSequenceNr
-
-    val maxPnr = partitionNr(firstJournalSequenceNr)
+    val maxPnr = partitionNr(atomicWrites.last.payload.last.sequenceNr)
     val firstSeq = atomicWrites.head.payload.head.sequenceNr
-    val minPnr = partitionNr(lastJournalSequenceNr)
+    val minPnr = partitionNr(firstSeq)
     val persistenceId: String = atomicWrites.head.persistenceId
     val all = atomicWrites.flatMap(_.payload)
 
@@ -104,28 +95,49 @@ class CassandraJournal extends AsyncWriteJournal with CassandraRecovery with Cas
     require(maxPnr - minPnr <= 1, "Do not support AtomicWrites that span 3 partitions. Keep AtomicWrites <= max partition size.")
 
     val writes: Seq[BoundStatement] = all.map { m =>
-      preparedWriteMessage.bind(journalId, maxPnr: JLong, m.journaSequenceNr: JLong, persistenceId, m.sequenceNr: JLong, m.serialized)
+      preparedWriteMessage.bind(persistenceId, maxPnr: JLong, m.sequenceNr: JLong, m.serialized)
     }
     // in case we skip an entire partition we want to make sure the empty partition has in in-use flag so scans
     // keep going when they encounter it
-    if (partitionNew(firstJournalSequenceNr) && minPnr != maxPnr) writes :+ preparedWriteInUse.bind(journalId, minPnr: JLong)
+    if (partitionNew(firstSeq) && minPnr != maxPnr) writes :+ preparedWriteInUse.bind(persistenceId, minPnr: JLong)
     else writes
+
   }
 
-  // TODO: FIX
-  override def asyncDeleteMessagesTo(persistenceId: String, toSequenceNr: Long): Future[Unit] =
-    Future(())
+  def asyncDeleteMessagesTo(persistenceId: String, toSequenceNr: Long): Future[Unit] = {
+    val logicalDelete = session.executeAsync(preparedInsertDeletedTo.bind(persistenceId, toSequenceNr: JLong))
 
-  // TODO: FIX
-  override def asyncReadHighestSequenceNr(
-    persistenceId: String,
-    fromSequenceNr: Long): Future[Long] = Future(1l)
+    val fromSequenceNr = readLowestSequenceNr(persistenceId, 1L)
+    val lowestPartition = partitionNr(fromSequenceNr)
+    val highestPartition = partitionNr(toSequenceNr) + 1 // may have been moved to the next partition
+    val partitionInfos = (lowestPartition to highestPartition).map(partitionInfo(persistenceId, _, toSequenceNr))
 
-  override def asyncReplayMessages(
-    persistenceId: String,
-    fromSequenceNr: Long,
-    toSequenceNr: Long,
-    max: Long)(recoveryCallback: (PersistentRepr) => Unit): Future[Unit] = Future(())
+    partitionInfos.map( future => future.flatMap( pi => {
+      Future.sequence((pi.minSequenceNr to pi.maxSequenceNr).grouped(config.maxMessageBatchSize).map { group => {
+          val delete = asyncDeleteMessages(pi.partitionNr, group map (MessageId(persistenceId, _)))
+          delete.onFailure {
+            case e => log.warning(s"Unable to complete deletes for persistence id ${persistenceId}, toSequenceNr ${toSequenceNr}. The plugin will continue to function correctly but you will need to manually delete the old messages.", e)
+          }
+          delete
+        }
+      })
+    }))
+
+    logicalDelete.map(_ => ())
+  }
+
+  private def partitionInfo(persistenceId: String, partitionNr: Long, maxSequenceNr: Long): Future[PartitionInfo] = {
+    session.executeAsync(preparedSelectHighestSequenceNr.bind(persistenceId, partitionNr: JLong))
+      .map(rs => Option(rs.one()))
+      .map(row => row.map(s => PartitionInfo(partitionNr, minSequenceNr(partitionNr), min(s.getLong("sequence_nr"), maxSequenceNr)))
+        .getOrElse(PartitionInfo(partitionNr, minSequenceNr(partitionNr), -1)))
+  }
+
+  private def asyncDeleteMessages(partitionNr: Long, messageIds: Seq[MessageId]): Future[Unit] = executeBatch({ batch =>
+    messageIds.foreach { mid =>
+      batch.add(preparedDeletePermanent.bind(mid.persistenceId, partitionNr: JLong, mid.sequenceNr: JLong))
+    }
+  }, Some(config.deleteRetries))
 
   private def executeBatch(body: BatchStatement ⇒ Unit, retries: Option[Int] = None): Future[Unit] = {
     val batch = new BatchStatement().setConsistencyLevel(writeConsistency).asInstanceOf[BatchStatement]
@@ -156,7 +168,7 @@ class CassandraJournal extends AsyncWriteJournal with CassandraRecovery with Cas
   }
 
   private case class SerializedAtomicWrite(persistenceId: String, payload: Seq[Serialized])
-  private case class Serialized(journaSequenceNr: Long, sequenceNr: Long, serialized: ByteBuffer)
+  private case class Serialized(sequenceNr: Long, serialized: ByteBuffer)
   private case class PartitionInfo(partitionNr: Long, minSequenceNr: Long, maxSequenceNr: Long)
 }
 
@@ -169,4 +181,4 @@ class FixedRetryPolicy(number: Int) extends RetryPolicy {
     if (nbRetry < number) RetryDecision.retry(cl) else RetryDecision.rethrow()
   }
 }
-
+*/
